+++ app/views/tops/index.html.erb
@@ -1,7 +1,7 @@
 <div class="top_image items-center text-center justify-center font-hannari">
   <div class="pb-16 pt-10 md:pt-0 lg:pt-0">
     <video class="w-full" autoplay loop muted playsinline>
-      <source src="<%= asset_path('top.mp4') %>" type="video/mp4">
+      <source src="<%= asset_path('top.webm') %>" type="video/mp4">
       お使いのブラウザはビデオタグをサポートしていません。
     </video>
   </div>
